+++ assets/javascripts/kitten/karl/pages/manifesto/index.js
@@ -1,0 +1,21 @@
+import React from 'react'
+import Hero from './components/hero'
+import Engagements from './components/engagements'
+import Discrimination from './components/discrimination'
+import Protect from './components/protect'
+import Defense from './components/defense'
+import UnderHero from './components/under-hero'
+
+
+const Manifesto = () => (
+  <>
+    <Hero />
+    <UnderHero />
+    <Engagements />
+    <Discrimination />
+    <Protect />
+    <Defense />
+  </>
+)
+
+export default Manifesto